LBAAS-838: add rate_limit_expunge_seconds config; run expunge hourly

+++ libra/admin_api/expunge/expunge.py
@@ -27,10 +27,8 @@ LOG = log.getLogger(__name__)
 class ExpungeScheduler(object):
     def __init__(self):
         self.expunge_timer = None
-        self.expire_days = cfg.CONF['admin_api']['expire_days']
-        if not self.expire_days:
-            LOG.info('Expunge not configured, disabled')
-            return
+        self.expire_lb_days = cfg.CONF['admin_api']['expire_days']
+        self.expire_rate_limit_secs = cfg.CONF['admin_api']['rate_limit_expunge_seconds']
         self.server_id = cfg.CONF['admin_api']['server_id']
         self.number_of_servers = cfg.CONF['admin_api']['number_of_servers']
         self.run_expunge()
@@ -40,37 +38,55 @@ class ExpungeScheduler(object):
             self.expunge_timer.cancel()
 
     def run_expunge(self):
-        day = datetime.now().day
-        if self.server_id != day % self.number_of_servers:
-            LOG.info('Not our turn to run expunge check, sleeping')
+        hour = datetime.now().hour
+        if self.server_id != hour % self.number_of_servers:
+            LOG.info('Not our turn to run expunge check, sleeping 1 hour')
             self.expunge_timer = threading.Timer(
-                24 * 60 * 60, self.run_expunge, ()
+                60 * 60, self.run_expunge, ()
             )
         with db_session() as session:
             try:
-                exp = datetime.now() - timedelta(
-                    days=int(self.expire_days)
-                )
-                exp_time = exp.strftime('%Y-%m-%d %H:%M:%S')
-                LOG.info(
-                    'Expunging deleted loadbalancers older than {0}'
-                    .format(exp_time)
-                )
-                count = session.query(
-                    LoadBalancer.status
-                ).filter(LoadBalancer.updated < exp_time).\
-                    filter(LoadBalancer.status == 'DELETED').delete()
-                counter = session.query(Counters).\
-                    filter(Counters.name == 'loadbalancers_expunged').first()
-                counter.value += count
-                session.commit()
-                LOG.info(
-                    '{0} deleted load balancers expunged'.format(count)
-                )
-                # TODO delete rate_limited_actions rows older than
-                # rate_limit_expunge_age seconds
+                if self.expire_lb_days:
+                    exp = datetime.now() - timedelta(
+                        days=int(self.expire_lb_days)
+                    )
+                    exp_time = exp.strftime('%Y-%m-%d %H:%M:%S')
+                    LOG.info(
+                        'Expunging deleted loadbalancers older than {0}'
+                        .format(exp_time)
+                    )
+                    count = session.query(
+                        LoadBalancer.status
+                    ).filter(LoadBalancer.updated < exp_time).\
+                        filter(LoadBalancer.status == 'DELETED').delete()
+                    counter = session.query(Counters).\
+                        filter(Counters.name == 'loadbalancers_expunged').first()
+                    counter.value += count
+                    session.commit()
+                    LOG.info(
+                        '{0} deleted load balancers expunged'.format(count)
+                    )
+                if self.expire_rate_limit_secs:
+                    exp = datetime.now() - timedelta(
+                        seconds=int(self.expire_rate_limit_secs)
+                    )
+                    exp_time = exp.strftime('%Y-%m-%d %H:%M:%S')
+                    LOG.info(
+                        'Expunging rate_limited_actions older than {0}'
+                        .format(exp_time)
+                    )
+                    count = session.query(
+                        RateLimitedActions
+                    ).filter(RateLimitedActions.use_time < exp_time).delete()
+                    counter = session.query(Counters).\
+                        filter(Counters.name == 'rate_limited_expunged').first()
+                    counter.value += count
+                    session.commit()
+                    LOG.info(
+                        '{0} old rate_limited_actions expunged'.format(count)
+                    )
             except:
                 LOG.exception('Exception occurred during expunge')
-        LOG.info('Expunge thread sleeping for 24 hours')
+        LOG.info('Expunge thread sleeping for 1 hour')
         self.expunge_timer = threading.Timer(
-            24 * 60 * 60, self.run_expunge, ())
+            60 * 60, self.run_expunge, ())
